# services/gateway/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for querying GPU telemetry data from an AI cluster.",
        "title": "GPU Telemetry API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "aravindgpd@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/gpus": {
            "get": {
                "description": "Returns the registered GPUs. Filter by model_name or hostname; paginate with limit/offset.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gpus"
                ],
                "summary": "List GPUs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter to one model. Case-insensitive substring match — 'h100' matches 'NVIDIA H100 80GB HBM3'. See /models for what's available.",
                        "name": "model_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to one host. Case-insensitive substring match.",
                        "name": "hostname",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum records to return (default 100, max 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.GPU"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/gpus/{id}/telemetry": {
            "get": {
                "description": "Returns DCGM metric samples for the specified GPU UUID, ordered by sample_at ascending. Optionally filter by metric_name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gpus"
                ],
                "summary": "Get GPU telemetry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GPU UUID (e.g. GPU-5fd4f087-86f3-7a43-...)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "DCGM metric filter (e.g. DCGM_FI_DEV_GPU_UTIL). Omit for all metrics.",
                        "name": "metric_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive lower bound on sample_at (RFC3339, e.g. 2025-07-18T20:42:34Z)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive upper bound on sample_at (RFC3339)",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum records to return (default 100, max 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.TelemetryRecord"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Liveness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/models": {
            "get": {
                "description": "Discovery endpoint — lists every distinct GPU model_name in the database with a count of GPUs running it. Use the model_name from this list as the `?model_name=` filter on /gpus and /telemetry.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gpus"
                ],
                "summary": "List GPU models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.ModelSummary"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/readyz": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Readiness probe",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        },
        "/telemetry": {
            "get": {
                "description": "Cross-GPU telemetry search. All filters are optional; combine freely.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "telemetry"
                ],
                "summary": "Query telemetry across GPUs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DCGM metric filter (e.g. DCGM_FI_DEV_GPU_UTIL)",
                        "name": "metric_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "GPU model filter. Case-insensitive substring match — 'h100' is enough; the full 'NVIDIA H100 80GB HBM3' also works. See /models for what's available.",
                        "name": "model_name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to one specific GPU UUID",
                        "name": "uuid",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive lower bound on sample_at (RFC3339)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Inclusive upper bound on sample_at (RFC3339)",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum records to return (default 100, max 1000)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.TelemetryRecord"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "store.GPU": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "device": {
                    "type": "string"
                },
                "gpu_index": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "model_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "store.ModelSummary": {
            "type": "object",
            "properties": {
                "gpu_count": {
                    "type": "integer"
                },
                "model_name": {
                    "type": "string"
                }
            }
        },
        "store.TelemetryRecord": {
            "type": "object",
            "properties": {
                "container": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ingested_at": {
                    "type": "string"
                },
                "labels_raw": {
                    "type": "string"
                },
                "metric_name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "pod": {
                    "type": "string"
                },
                "sample_at": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    }
}